@@ -6,12 +6,14 @@
   "scripts": {
     "build": "rimraf ./dist && tsc",
     "dev": "tsx watch src/index.ts",
-    "start": "node dist/index.js",
-    "restart": "npm run build && npm run start",
+    "lint": "eslint ./src/",
+    "lint:fix": "eslint ./src/ --fix",
+    "migration:drop": "typeorm-ts-node-esm schema:drop -d src/infrastructure/db/database.ts",
     "migration:gen": "typeorm-ts-node-esm migration:generate -d src/infrastructure/db/database.ts src/infrastructure/db/migrations/migration",
-    "migration:sync": "typeorm-ts-node-esm migration:run -d src/infrastructure/db/database.ts",
     "migration:revert": "typeorm-ts-node-esm migration:revert -d src/infrastructure/db/database.ts",
-    "migration:drop": "typeorm-ts-node-esm schema:drop -d src/infrastructure/db/database.ts"
+    "migration:sync": "typeorm-ts-node-esm migration:run -d src/infrastructure/db/database.ts",
+    "restart": "npm run build && npm run start",
+    "start": "node dist/index.js"
   },
   "keywords": [],
   "author": "",
@@ -29,14 +31,16 @@
     "typeorm": "^0.3.21"
   },
   "devDependencies": {
+    "@eslint/js": "^9.25.1",
     "@types/bcryptjs": "^2.4.6",
     "@types/express": "^5.0.0",
     "@types/jsonwebtoken": "^9.0.9",
-    "eslint": "^9.22.0",
+    "eslint": "^9.25.1",
     "jest": "^29.7.0",
     "rimraf": "^6.0.1",
     "ts-node-dev": "^2.0.0",
     "tsx": "^4.19.3",
-    "typescript": "^5.8.2"
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.31.0"
   }
 }
