@@ -8,37 +8,37 @@ import { UserRole } from "../../core/entities/user.entity.js";
 
 const UserRouter = express.Router();
 
-const userRepository = container.resolve<UserRepository>('UserRepository');
-const userService = new UserService(userRepository);
-const userController = new UserController(userService);
+const user_repository = container.resolve<UserRepository>('UserRepository');
+const user_service = new UserService(user_repository);
+const user_controller = new UserController(user_service);
 
-UserRouter.post('/auth/register', userController.createUser.bind(userController));
-UserRouter.post('/auth/login', userController.login.bind(userController));
+UserRouter.post('/auth/register', user_controller.createUser.bind(user_controller));
+UserRouter.post('/auth/login', user_controller.login.bind(user_controller));
 
 // Admin routes
 UserRouter.post(
     '/users', 
     authMiddleware,
     isAuthorized([UserRole.ADMIN]),
-    userController.createUser.bind(userController)
+    user_controller.createUser.bind(user_controller)
 );
 UserRouter.put(
     '/user/:id', 
     authMiddleware,
     isAuthorized([UserRole.ADMIN]),
-    userController.updateUser.bind(userController)
+    user_controller.updateUser.bind(user_controller)
 );
 UserRouter.get(
     '/users/role/:role', 
     authMiddleware,
     isAuthorized([UserRole.ADMIN]),
-    userController.getAllUsersByRole.bind(userController)
+    user_controller.getAllUsersByRole.bind(user_controller)
 );
 UserRouter.delete(
     '/user/:id', 
     authMiddleware, 
     isAuthorized([UserRole.ADMIN]), 
-    userController.deleteUser.bind(userController)
+    user_controller.deleteUser.bind(user_controller)
 );
 
 // User routes
@@ -46,13 +46,13 @@ UserRouter.get(
     '/users/me', 
     authMiddleware, 
     isAuthorized([UserRole.USER]), 
-    userController.getUserProfile.bind(userController)
+    user_controller.getUserProfile.bind(user_controller)
 );
 UserRouter.put(
     '/users/me', 
     authMiddleware, 
     isAuthorized([UserRole.USER]), 
-    userController.updateUserProfile.bind(userController)
+    user_controller.updateUserProfile.bind(user_controller)
 );
 
 // Support and admin routes
@@ -60,19 +60,19 @@ UserRouter.get(
     '/user/email/:email', 
     authMiddleware, 
     isAuthorized([UserRole.ADMIN, UserRole.SUPPORT]), 
-    userController.getUserByEmail.bind(userController)
+    user_controller.getUserByEmail.bind(user_controller)
 );
 UserRouter.get(
     '/user/:id', 
     authMiddleware, 
     isAuthorized([UserRole.ADMIN, UserRole.SUPPORT]), 
-    userController.getUserById.bind(userController)
+    user_controller.getUserById.bind(user_controller)
 );
 UserRouter.get(
     '/users', 
     authMiddleware, 
     isAuthorized([UserRole.ADMIN, UserRole.SUPPORT]), 
-    userController.getAllUsers.bind(userController)
+    user_controller.getAllUsers.bind(user_controller)
 );
 
 export default UserRouter;
